feat: implement dynamic import
